Migrate legacy test_pointwise.py; Skip legacy test_normalization.py
