+++ tests/python/direct/README.md
@@ -40,6 +40,7 @@ The `tests/python/direct` directory contains the following test files:
 - test_matmul.py
 - test_nan.py
 - test_optimization_passes.py
+- test_pointwise.py
 - test_python_frontend.py
 - test_repro.py
 - test_sdpa.py
@@ -47,11 +48,10 @@ The `tests/python/direct` directory contains the following test files:
 #### Remaining legacy test files
 - test_inline_definitions.py --- Inline definitions is not supported in direct bindings.
 - test_narrow_precision.py
-- test_normalization.py
-- test_pointwise.py
 
 #### Legacy-Only test files
 - test_moe.py --- Uses neither legacy nor direct nvfuser python frontend
+- test_normalization.py --- InstanceNorm3dNVFuser is defined in nvfuser.contrib.nn.normalization
 - test_schedule_ops.py --- Scheduling API will be different in direct bindings
 
 ## test_matmul.py
@@ -98,6 +98,12 @@ The following tests exist in `tests/python/direct/test_python_frontend.py` but a
 - `test_define_noncontiguous_tensor` - corresponds with `TestDefine::test_noncontiguous`
 - `test_define_broadcast_tensor` - corresponds with `TestDefine::test_broadcast`
 - `test_define_tensor_contiguity_with_stride_order` - corresponds with `TestDefine::test_contiguity_with_stride_order`
+
+### Migrated tests in `tests/python/direct/test_python_frontend.py` from `tests/python/test_pointwise.py`
+- `test_cpu_add`
+- `test_full_with_cpu_inputs`
+- `test_input_forwarding_device`
+- `test_single_segment_multi_device`
 
 ### Migrated tests in `tests/python/direct/test_python_frontend.py` from `tests/python/test_nan.py`
 - `test_validate_precomputed_values`
@@ -223,11 +229,18 @@ The following 19 issue-specific tests have been migrated from the main frontend 
 - `test_issue2275_repro2` - Tests fix for issue 2275 (unpadded concatenation operations with trigonometric functions); Maps to legacy `test_unpadded_catop_issue2275_repro2`
 - `test_issue2317` - Tests fix for issue 2317 (reduction transpose scheduling); Maps to legacy `test_reduction_transpose_sched_issue2317`
 - `test_issue2354` - Renamed from legacy `test_matmul_issue_2354` in `test_matmul.py`
+- `test_issue2395` --- Migrated from `test_issue_2395` in legacy test_pointwise.py
 - `test_issue2532` - Renamed from legacy `test_repro_issue2532` in `test_matmul.py`
 - `test_issue2545` - Tests fix for issue 2545 (complex operations with empty tensors and concatenation); Maps to legacy `test_remove_empty_issue_2545`
 - `test_issue2549` - Tests fix for issue 2549 (broadcast_in_dim and division operations); Maps to `test_fix_2549`
+- `test_issue2664_repro1` --- Migrated from `test_inplace_issue2664` in legacy test_pointwise.py
+- `test_issue2664_repro2` --- Migrated from `test_inplace_post_bcast` in legacy test_pointwise.py
+- `test_issue2664_repro3` --- Migrated from `test_multi_inplace` in legacy test_pointwise.py
+- `test_issue2664_repro4` --- Migrated from `test_implicit_bcast_inplace` in legacy test_pointwise.py
 - `test_issue2755` - Tests fix for issue 2755 (slice operations with negation)
+- `test_issue2853` --- Migrated from legacy test_pointwise.py
 - `test_issue3192` - Mapped from legacy `test_squeeze_issue_3192` in `test_alias.py`
+- `test_issue3227` --- Migrated from `test_bcast_different_extent` in legacy test_pointwise.py
 - `test_issue3369` - Mapped from legacy `test_square_linear` in `test_optimization_passes.py`
 - `test_issue3292` - Tests fix for issue 3292 (complex tensor operations with manual normalization and padding)
